Add conversion impl for chemrust-core
