+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "castep-cell-io"
-version = "0.2.10"
+version = "0.2.11"
 edition = "2021"
 author = ["TonyWu20"]
 description = "A crate helping to parse, edit and save `castep` input file format `.cell`"
