+++ src/cell_document/chemrust_impl.rs
@@ -15,29 +15,28 @@ use super::CellEssentials;
 use super::LatticeCart;
 
 pub fn to_cell_document<T: CrystalModel>(model: &T) -> CellDocument {
-    let lattice_bases = model.get_cell_parameters().lattice_bases();
-    let (a, b, c) = (
-        lattice_bases.column(0).into(),
-        lattice_bases.column(1).into(),
-        lattice_bases.column(2).into(),
+    let lattice_param_block = LatticeParamBlock::from(model.get_cell_parameters());
+    let ionic_pos_block = IonicPositionBlock::from_atom_data(
+        model.get_atom_data(),
+        &model.get_cell_parameters().lattice_bases(),
     );
-    let lattice_cart = LatticeCart::new(a, b, c);
-    let lattice_param = LatticeParam::LatticeCart(lattice_cart);
-    let lattice_param_block = LatticeParamBlock::new(LengthUnit::Ang, lattice_param);
-    let atom_data = model.get_atom_data();
-    let symbols = atom_data.symbols_repr();
-    let coords = atom_data
-        .coords_repr()
-        .iter()
-        .map(|cd| cd.cart_to_frac(&lattice_bases).raw_data().into())
-        .collect::<Vec<[f64; 3]>>();
-    let ionic_positions = symbols
-        .iter()
-        .zip(coords.iter())
-        .map(|(&symbol, &coord)| IonicPosition::new(symbol, coord, None))
-        .collect::<Vec<IonicPosition>>();
-    let ionic_pos_block =
-        IonicPositionBlock::new(LengthUnit::Ang, ionic_positions, POSITIONS_FRAC, true);
     let model_description = CellEssentials::new(lattice_param_block, ionic_pos_block);
     CellDocument::new(model_description)
+}
+
+impl<T> From<&T> for LatticeParamBlock
+where
+    T: UnitCellParameters,
+{
+    fn from(value: &T) -> Self {
+        let lattice_bases = value.lattice_bases();
+        let (a, b, c) = (
+            lattice_bases.column(0).into(),
+            lattice_bases.column(1).into(),
+            lattice_bases.column(2).into(),
+        );
+        let lattice_cart = LatticeCart::new(a, b, c);
+        let lattice_param = LatticeParam::LatticeCart(lattice_cart);
+        LatticeParamBlock::new(LengthUnit::Ang, lattice_param)
+    }
 }
